Add player interaction implementation
PlayerInteraction is used by the player in order to use interactable
objects. Using a raycast from the scene camera (which should be used for
first person), it will look for an object with the 'Interactable' tag,
and highlight it while it can be interacted with. The player can then
use left click to interact with the object, calling Use() on the
InteractionBase script

The scripts have been tested in the sandbox scene

InteractionBase will now use the default material for highlighting if
not set in the inspector, preventing the object from turning pink when
not set