Add core dialogue system
The core system revolves around the DialogueManager and DialogueGraph.
Only one DialogueManager script should exist in a scene, and is accessed
using its static instance. It is passed a dialogue scene graph, and will
'execute' it, displaying its dialogue.

The DialogueGraph contains the bulk of the logic, but dialogue is to be
implemented with the DialogueSceneGraph script, instanciated on a game
object. For further explanation on its use, refer to TDD, or ask Thomas.