Add test dialogue implementation, fix warnings
The sandbox contains a test dialogue implementation for testing, with an
interaction starting it. To do this, a new trigger component has been
created: StartDialogueComponent.

A warning would show up every time a new node was created due to the way
GetValue was handled. This has been fixed by making it always return
null in BaseNode.