Add core sitting system, Implement in sandbox
The sitting system uses a new controller script, along with a SeatInfo
script that is placed on a seat object. The controller script will move
the player to the 'sitting' position with a lerp, then limmit rotation
in both X and Y. On pressing space, the player is lerped to the
'standing' position next to the seat.

Sitting should be done through an interaction using SitComponent, under
Player Controllers/Sitting Scripts.