Fix TransitionManager, Complete working test
TransitionManager can now correctly transition between scenes and update
the player's position (controller swapping untested currently). A new
trigger component has also been created to trigger a scene transition,
called ChangeSceneComponent.

For scene tranitions to work, a 'base' scene containing the player and
the TransitionManager script needs to be the first loaded scene, with
the 'content' scene as the second. Only 2 scenes can be loaded at any
time.

The last thing to add is an animation to hide the transition, but as of
now, it is unnoticable due to the speed of loading the demo scenes.