Add transitional sitting and collection objects
When transitioning into a scene using the sitting controller, it will
not play an animation, and just make the player sit down in the nearest
seat.

A collection can now have a game object/s become enabled in the train
scene after collecting all the items. A 'CollectedObject' script will
need to be created for each collection, using a static bool to set
active on the attached game object. The collection manager will set the
bool after all items are collected.