Add help system
The help system only uses two scripts: a trigger component,
ProgressionItemComponent, and a manager script, HelpMessageManager.

The trigger component is used on items tied to progression, and set a
help message and how long to wait before showing it. The manager is used
by these trigger components, and it waits until the time is up before
sending the graph to DialogueManager.