Fix game for build
TrainCarrage and TT_Level_Japan_v1 have been fixed to allow scene
transitions to work using the new system. Currently, a sphere behind the
player is sued to transition.

The the carrage, the first booth on the right can be sat in.

'NewBehaviourScript' is a temp script to be able to load 2 scenes on
start up.