Implement paining on color ball hit
When a color ball collides with an object with a PaintableScript, it
will put its splat texture on the object with the selected color.

For an object to be paintable, it needs to have a PaintableScript and a
material with DecalShader. The size of the decal is dependant on the
PaintableScript's 'textureSize', with a larger value providing a smaller
decal (as it is more compact).