Add poder pouring minigame
When the player has the 'spoon' over the powder gather trigger, a bar
will show, and the player needs to click while the indicator is in the
'sweet spot'. After this, the spoon needs to enter the powder deposit to
drop off the powder. After doing this a set number of times, the
minigame is won (currently a debug log).