Add mesh water script, implement in japan scene
The script applies scrolling perlin noise to the height of a meshes verticies
to create a wave effect.

The cube in the japan scene for the water has been replaced with a plane
with the script on it, however the plane has very few verticies currently,
so the effect isnt noticable.