Change fade transition to blinking
Instead of having a black image fade in and out, now two black images
move from the edges toward the center to cover the screen.

The two images have a fading gradiant, but the shader used for gradients
previously didnt allow transparency, so it has been modified to do so.