Fix blink transition, add blink to cup transition
The fading bars using in the blink transition worked in the editor, but
had issues in build. This problem had to do with rotation and scaling,
and has been fixed with a modification to the gradiant shader.

The cup animation now has the cup raised, then a blink transition is used.